Shifted inventory indexes into separate folder
